+++ doc/source/doxygen-docs/changelog.md
@@ -11,6 +11,7 @@
     - New macro ASSERT_NEAR_(). Defined new macros with correct English names ASSERT_LT_(), etc. deprecating the former ones.
   - \ref mrpt_math_grp
     - New static methods with semantic-rich names: mrpt::math::TPlane::From3Points(), mrpt::math::TPlane::FromPointAndLine(), ...
+    - New asString() methods in mrpt::math::TPlane, mrpt::math::TLine2D, mrpt::math::TLine3D
 - Deprecations:
   - mrpt::system::TParameters is now deprecated.
 - BUG FIXES: